class for drawing objects

+++ AbstractObject.h
@@ -6,7 +6,7 @@ class AbstractObject
 protected:
 	Vector2 position;
 	Vector2 size;
-	int velocity;
+	float velocity;
 
 public:
 	Vector2 getPosition();
